+++ experimental/packages/opentelemetry-sdk-metrics-base/test/util.ts
@@ -29,6 +29,7 @@ import { Measurement } from '../src/Measurement';
 import { isNotNullish } from '../src/utils';
 import { HrTime } from '@opentelemetry/api';
 import { Histogram } from '../src/aggregator/types';
+import { ObservableCallback } from '@opentelemetry/api-metrics';
 
 export const defaultResource = new Resource({
   resourceKey: 'my-resource',
@@ -127,5 +128,18 @@ export function assertPartialDeepStrictEqual<T>(actual: unknown, expected: T, me
   const ownNames = Object.getOwnPropertyNames(expected);
   for (const ownName of ownNames) {
     assert.deepStrictEqual((actual as any)[ownName], (expected as any)[ownName], `${ownName} not equals: ${message ?? '<no-message>'}`);
+  }
+}
+
+export class ObservableCallbackDelegate {
+  private _delegate?: ObservableCallback;
+  setDelegate(delegate: ObservableCallback) {
+    this._delegate = delegate;
+  }
+
+  getCallback(): ObservableCallback {
+    return observableResult => {
+      return this._delegate?.(observableResult);
+    };
   }
 }